build app outside of React
Able to launch app with Electron and initiate ui with Ctr+Alt+P.
Able to draw on canvas and save image to a file, specifying filepath
in dialog window.

The client side app is now built outside of React in src/index.js.
This is preventing the initial screenshot image from loading in the
Canvas.js constructor.