BUG: screen capture not working in packaged version
Using screenshot-desktop to capture image, a promise that returnes an
image buffer array. Canvas component handles formating the buffer array
to correct Uint8Array format and loads image with data url. This works
in developent, but not in packaged app.

I was not able to se console logs made in electron.js so I created a
Logger class that stores all logs and then sends them to the render
process once it is ready. This allows the BrowserWindow devtools console
to display logs that were only visible from the cmd line console.